code refractor and queue position added

+++ backend/src/main/java/com/livedesk/ticket/dto/CreateTicketResponse.java
@@ -5,5 +5,5 @@ import java.util.UUID;
 public record CreateTicketResponse(
         UUID ticketId,
         String sessionToken,
-        Integer queuePosition
+        Long queuePosition
 ) {}
